Move html report generation to separate file
